standardise ajax loader (code health)
I was getting a 404 due to a bad url in mf-common.less with
the generic ajax loader. This moves the image there and changes
all references to it. In process standardised the name of the
class that provides the ajax loader to 'loading' applied to watchlist
loading star and edit screen as well as for dynamically loaded pages

Change-Id: I2c679aa93f1e1313966bb4b9ab63da23a82e7320