Update patch set 1
Patch Set 1:

The only thing that worries me is that this hook is in a file where all the general hooks are, but unlike the rest of them is very specific, i.e. for watchlist functionality only. I can imagine this function becoming a function with 10 responsibilities in future.

We can leave it as it is now I guess, and if we get more special cases like this then think about refactoring. I don't see any better place for it right now anyway (I guess SpecialMobileWatchlist.php is not the right place either).

Patch-set: 1