Add View, an abstraction over jQuery elements.
This might help us untangle the JavaScript code a bit. Inspired by
Backbone's views and Knockout.js's ViewModels.

See docs in the code and tests for more details. Real use case example
in subsequent commits.

It's the first version, we can change it or rewrite it later on
when we discover our use patterns for it.

Also adds Hogan as a templating engine.

Change-Id: Ic70e708328e73d07c82f767d24be5299d94b9699