Update patch set 1
Patch Set 1: There's a problem with this change, please improve

Use array_merge() instead of '+' operator. '+' works, but it is uncommon and has potentially unexpected consequences. For instance, the '+' operator gives preference to the left-most array when both arrays have matching keys. The way this is written means that you would not be able to overwrite keys of $wgMobileFrontendResourceTemplate, which I think makes this change much less convenient than it could be.

Patch-set: 1