Update patch set 1
Patch Set 1:

Also, I realize you might not be into the whole brevity thing, el duderino, but it would be nice if we standardized our global variable names. Most of the MobileFrontend-specific global vars are named $wgMF<whatever>. I encourage you to rename $wgMobileFrontendResourceTemplate to something like $wgMFResourceTemplate. Also, the contents of that var are less of a template and more of a base, or boilerplate. Consider making the name more clear.

Patch-set: 1