Update patch set 1
Patch Set 1:

It's because $this->mobileView in MobileContext is already set before $this->forceMobileView is set. Which means that MobileContext::shouldDisplayMobileView() returns the already set mobileView property rather than reevaluating shouldDisplayMobileView().

A quick glance through the logic suggests to me that we may actually want to rethink how some of that is currently working!

Patch-set: 1