Update patch set 2
Patch Set 2: There's a problem with this change, please improve

I think it's generally good, but I have two comments:

1. I used emit() instead of trigger() in View (https://gerrit.wikimedia.org/r/gitweb?p=mediawiki/extensions/MobileFrontend.git;a=blob;f=javascripts/common/mf-view.js;h=f2cdd883e282888c12d07b2ba612d1544ae64cc3;hb=HEAD) because trigger just sounds DOM-specific. This is personal preference so I can change it. We should decide on one name though.

2. I'd change the jQuery's API a bit instead so that it doesn't pass the event object to on() callbacks (we won't ever use this object since those are not DOM events). Also, I'd fix one inconsistency in jQuery: trigger() accepts an array as arguments which are later passed to on() callbacks, but they are not passed as an array. What I did in View is:

.on('someEvent', function(arg1, arg2) {});
.emit('someEvent', arg1, arg2);

instead of:

.on('someEvent', function(ev, arg1, arg2) {});
.trigger('someEvent', [arg1, arg2]);

Changing this behaviour is also a small reason for changing trigger() to emit().

Ideally, View's functions and the one we use for main events would be the same, but we can take care of it later, when https://bugzilla.wikimedia.org/show_bug.cgi?id=44264 is resolved.

Patch-set: 2