Update patch set 1
Patch Set 1: Do not submit

I think the team needs to consense on whether or not it's a good idea to add this as a precommit hook. This creates yet another dependency that anyone who wants to commit to our project will need to install/setup. This would be good to discuss with the rest of the team and try to come to decision about. Marking -2 til this gets discussed more broadly.

Patch-set: 1