Update patch set 1
Patch Set 1: Do not submit

As discussed with Julisuz let's fix bug https://bugzilla.wikimedia.org/show_bug.cgi?id=43271 rather than piling on more hacks.

This is not in stable yet and only effects a small handful of pages so I'd be keen to leave this for the time being.

Patch-set: 1