Update patch set 1
Patch Set 1:

The way the pre-commit hooks are currently done in MobileFrontend makes them all-or-nothing. Maybe the answer is to break the components of this file out into individual scripts to make it easier to optionally deal with. That way support for codesniffer can be added, and we can all enable what we want without having to create some custom workaround that may drift from what's available in in scripts/pre-commit.

I -2'd because I believe most (if not all) of us on the development team are using the pre-commit hooks defined in this file sym-linked to .git/hooks/pre-commit, which if this were merged, would cause breakage for us until we all setup/installed codesniffer. And like I mentioned before, I do not think it makes sense to make this a requirement, unless the team generally agrees to it, and as a result I do not believe this patch should be submitted.

Patch-set: 1