Update patch set 3
Patch Set 3: There's a problem with this change, please improve

(4 inline comments)

I can't wait for this to get merged. I really want to get rid of the JS-but-no-jQuery mode.

But, I'm not sure if introducing new functionality to the RL is necessary (mobileTargets). Maybe I'm missing something, but it seems to me that we could simply use dependencies for this.

As an example let's take beta. Instead of adding 'mobileTargets' => array( 'beta' ) to mobile.beta.dependencies and mobile.beta.jquery, we could create a module called mobile.beta, which wouldn't contain any scripts or styles, but instead it would only depend on all the modules we want in beta:

$wgResourceModules['mobile.beta'] = array(
    'dependencies' => array(
        'mobile.stable',
        'mobile.beta.dependencies',
        'mobile.beta.jquery'
    )
);

Similarly, we would create mobile.stable and mobile.alpha. And then we would simply load one and only one module (mobile.stable, mobile.beta or mobile.alpha) and this module's dependencies would simply pull all the other modules required for the given mode (stable/beta/alpha) to work.

What do you think?

Patch-set: 3