Update patch set 1
Patch Set 1: There's a problem with this change, please improve

It juts for me after applying this patch. It's not a browser hack, it's more of a font hack. Depending on the font your device uses, the bold variant of the font might have a different default line-height than the normal variant. This is necessary to keep it consistent and avoid jutting.

Patch-set: 1