Update patch set 3
Patch Set 3:

after chatting more with Max in irc, it's clear the approach taken here with loading the template files is the same as RL uses to load css/js files, so my concerns around the file loading are probably less of a big deal. It still seems unnerving to me, but I'm willing to concede this one. I would still like us to consider being a little more defensive with this approach, but I'm not going to block on this.

Patch-set: 3