Update patch set 8
Patch Set 8: There's a problem with this change, please improve

Ok, Juliusz or I should amend this patchset with the styling changes now. Code is fine but I'd hate to merge it without that styling change as looks very odd like this!

Also it's not clear from the story whether this should be the title of the page or the prefix text. We should clarify with design/product!

Patch-set: 8