Update patch set 1
Patch Set 1: There's a problem with this change, please improve

Max I saw the bug you raised about watchlist stars on the mobile site but this takes away that functionality altogether as well as removing js thumbnails.

Note this is not such a high priority as the image thumbnails are limited to beta.

Patch-set: 1