Update patch set 1
Patch Set 1:

Actually, moving these changes into the template is rather difficult and will require some more significant changes. This could be done a little more easily with https://gerrit.wikimedia.org/r/48209. This is made extra difficult because the count needs to be added above the donate image button (which is dynamically added by JS) in a portion of the template that currently does not have a template var available (in the content_wrapper div before the content div). Even if there were, setting template variables from within the SpecialPage seems non-trivial at best (unless we get the change mentioned above merged along with the corresponding core change). I suggest we leave it as I initially wrote it for now.

Patch-set: 1