Update patch set 1
Patch Set 1: Code-Review-1

Why this it call this at all?

SkinTemplate and OutputPage provides two ways to do this:

A) headelement + printTail

* Set $this->useHeadElement = true;
* Use $out->headElement()
*  $this->printTrail()

B) headelement + bottomscripts

* Set $this->useHeadElement = true;
* Use $out->headElement()
* $this->bottomScripts()

C) get scripts and styles queue manually

* $out->getHeadLinks(), $out->buildCssLinks(), $out->getHeadScripts(), $out->getHeadItems()
*  $this->bottomScripts()

But what is the reason for going even lower and manually looping through the modules and stuff? Isn't the whole point of having Target in ResourceLoader in MediaWiki core that one can call the same methods and it will only queue the stuff targetted for mobile?

Patch-set: 1