Update patch set 2
Patch Set 2:

This fails in the following case in beta mode (dynamic loading):
* go to page A that's watched
* unwatch A
* click a link to page B
* click a link back to page A

You see the watchlist star indicating that it's still watched, as the cache variable hasn't changed.

We could clear the cached entry when doing a watch/unwatch operation, but that'll still fail in multiple-device situations (if I unwatch on another browser), however that's an edge case.


If you don't mind that being flaky in beta, this is still an improvement for regular mode.

Patch-set: 2