Update patch set 2
Patch Set 2:

OK the main reason I introduced this was for the situation where you click on uploads and can't upload and haven't uploaded before - as a result you get a page saying "Photos bring pages on JDLR to life. There are many pages which have no photos due to a severe lack of freely licensed photos. Help bring these pages to life by donating your images that you are willing to give away and could be of use to the wider JDLR editing community." but you can't upload which makes it look a bit useless...

Patch-set: 2