Update patch set 1
Patch Set 1:

what is extra weird is i can see 'read in another language' as a greyed out section before the sections fully load. once the sections have laoded and become active, 'read in another language' is nowhere to be seen when only one language link is present.

if there are > 1 language links on the page, this works just fine.

Patch-set: 1