(mingle #410) Add X-Analytics request/response header handling support
This change will cause MF to add analytics info to X-Analytics response
header when user is in alpha or beta, and/or if there is an X-Analytics
request header (in WMF production, this will happen if the user
is accessing the site via ZeroRatedMobileAccess). The X-Analytics
response header will only be set if the aforementioned criteria
are met, AND wgMFEnableXAnalyticsLogging (added in this changeset) is
set to true.

In the longer term, it would be good to move the X-Analytics response
header handling into core, perhaps into OutputPage - but since
Zero and MobileFrontend are the only components currently using
the X-Analytics analytics headers, it seemed prudent and more expedient
to start with the handling in MobileFrontend.

Relates to change: I14973cce2d2e80f480aabefa2dfbbfd5936db6cb
Change-Id: I93758fc76721d89f9c32a28c1821f3e59eb384ef