Update patch set 1
Patch Set 1:

Yeh I figured something like this was possible but couldn't work it out. I'm still not sure if this is the best approach but I couldn't think a better way of doing it.

Ideally we would be able to run a function on the special page object to see if it is written for mobile. Not sure if possible.
specialPage->isMobileOptimised()

Patch-set: 1