Update patch set 2
Patch Set 2:

More specifically, my question was 'what technology/technologies loads the CentralNotice banner?' If it's JS filling an empty div with an id of 'siteNotice', then this patchset looks fine, but if the banner is ever to be loaded via PHP, then this patch set does not look fine :)

I only bring this up because provisions are made for loading siteNotice via PHP in core. So, if the banner gets loaded from JS and we only care about targeting users who support Javascript, then I think this patchset is good to be merged.

Patch-set: 2