Update patch set 1
Patch Set 1: Verified+1 Code-Review+1

I was wondering what form of English is widely used and a grep of mediawiki suggests American spellings are more frequent than British English.

This is a debug message so it doesn't seem very important and happy to +2 but it would be nice to get an understanding of whether I (a British English speaker) and other developers should be writing code in British English or American English.

Siebrand ... ?

Patch-set: 1