Update patch set 1
Patch Set 1:

"stats cachedump" on mc1001 doesn't show a predominance of MF cache entries. It's about 7% of items in slab 142. Unlike in the MySQL parser cache, we have LRU eviction in memcached, which makes a high write rate more tolerable. So I'm not sure if this change would be beneficial.

Patch-set: 1