Update patch set 3
Patch Set 3: Code-Review-1

If anyone disagrees feel free to undo my -1 - I'm open to discussion :) Sadly -1s are the only way to get comments noticed!

I guess my problem is actually with the code setup isAlphaGroupMember is quite a complicated method - you'd expect it to simply be a lookup but it also sets a property on MobileContext.
checkUserStatus gets called every time a call to isAlphaGroupMember or isBetaGroupMember gets called which seems wrong and now we are now adding an analytics log item to that call as well. We call isAlphaGroupMember and isBetaGroupMember all over the place - it should be a cheap function.
I'd prefer if in SkinMobile.php (or some other entry point) We simply ran $context->logMobileMode() which calls isAlphaGroupMember / isBetaGroupMember and adds the header...

Patch-set: 3