Update patch set 3
Patch Set 3:

This is a more stylistic comment, but maybe we should rename 'X-CS' and not use it in the code. CS stands for Carrier Short and X-CS is a total arbitrary name for this header and since we are going to use it to capture different kinds of mobile usage we might as well come up with a better name.

Patch-set: 3