Update patch set 1
Patch Set 1: Code-Review-1

No but it can definitely be improved to solve the same goal.

An alternative way might be to actually document the init method and in the function notes actually mention the keys you are trying to get grep for. The words standalone are not useful javascript comments and do not help the code become more readable. This should be more important than grepping :)

This is an example of how we are currently documenting methods:
https://gerrit.wikimedia.org/r/gitweb?p=mediawiki/extensions/MobileFrontend.git;a=blob;f=javascripts/common/mf-view.js;h=de21571eaebc689defad63a5cde93240e02c4697;hb=HEAD#l92

Patch-set: 1