Update patch set 4
Patch Set 4:

The other problem with this is that resources loaded via ResourceLoader do not vary the cache by device like they do for MobileFrontend text. That means that once that device-specific variable is set for a given resource, it will be cached along with the rest of the resource.

Patch-set: 4