Update patch set 1
Patch Set 1: Code-Review-1

I do indeed get an upload failure if I select a GIF image... we can probably live with that but it'd be better to fix so that PNG screenshots etc can be uploaded.


With a JPEG... I get filenames like "File:Epic smiley 2013 03 18 13 05.jpg" which are hard to scan visually. I'd recommend using "-" to separate the date and time chunks like:

"File:Epic smiley 2013-03-018 13-05.jpg"

Patch-set: 1