Update patch set 2
Patch Set 2:

I too wanted to kill create,show and hide Overlay functions - but the views don't really work here. We need some sort of controller to manage all the overlays - ensure which one gets shown, which are open and when to return to article. Maybe we should sit down and work out how this might work?

Patch-set: 2