Update patch set 2
Patch Set 2:

As I read the code, I think this should work OK even with Varnish caching. The login link is generated from the article title, which will be consistent and unique for cached resources. That is, the 'returnto' value should always be the same for a given article, and therefore should not result in weirdness due to caching.

Patch-set: 2