Update patch set 2
Patch Set 2:

This'll limit the input area to accept only 240 Unicode characters; the limit for filename length however applies to UTF-8 bytes. (Story 347 code already deals with trimming the filename.)

Still a good guide for Latin/ASCII filenames, but could be misleading for other scripts where you might have lots of characters that you typed get trimmed.

On the other hand if we're trimming for the filename but passing through a description, there's no reason to trim or limit the description at all... should we actually apply a limit?

Patch-set: 2