Update patch set 3
Patch Set 3: Code-Review-1

jdlrobson it's pretty common to call parent::__construct(). in fact, you also need to invoke UnlistedSpecialpage::__construct() in UnlistedSpecialMobilePage.php. 

the approach i initially suggested is a little unconventional. i suggested it because i hadn't seen UnlistedSpecialMobilePage.php yet and figured what i was suggesting would give you more flexibility. but if you /always/ want UnlistedSpecialMobilePage::clearPageMargins() to run for every child of UnlistedSpecialMobilePage, then stick it in the constructor. if you want it to run in most invocations of UnlistedSpecialMobilePage but not necessarily all, my approach would make more sense.

Take a look at UnlistedSpecialPage class in includes/SpecialPage.php in core to see how the constructor works, and there are examples of invocations in the special pages in MobileFrontend.

Patch-set: 3