Regression: Remove OverlayManager, Fix photo uploads
a better way to do this is for Overlay's when created to specify a
parent. If a parent is specified closing the overlay will show the
parent overlay. If an overlay has no parent when hiding it will switch
from overlay mode to article mode.

This seems neater. Fix call to non-existant close to trigger hide.

Change-Id: Ie8dfe4d833c96b16eee05a84fcf6787782f18504