Update patch set 2
Patch Set 2:

(2 comments)

As far as I remember, Maryana and I decided to do it like this for a start. I could also show the overlay when the page loads after logging in I guess. We can change it later after more input on how to do it best.

Patch-set: 2