Update patch set 4
Patch Set 4:

(8 comments)

It functionally works for me on Chrome on my Android...
In terms of padding/margin I moaned about this to Vibha but she suggested optimising the design for iphone. She made some tweaks so it is no surprise if there are slight differences between this and mockups.

In terms of file naming conventions. I'm happy to have one but in the absence of an agreement we should just go ahead and merge. Filenames are easy to change in bulk later once we have agreement.

Patch-set: 4