Update patch set 2
Patch Set 2: Code-Review+1

I'm still not able to run it (because of the burnt laptop, should have a new working dev env later today), but it looks OK.

I'm not sure why we need to specify the number of expectations. If there's an error in the test function, it will fail anyway, so there's no way it will pass when some expectations in it pass and an error makes other expectations not run. Krinkle, can you point me to some explanation?

Patch-set: 2