Update patch set 1
Patch Set 1:

If the description changes and no longer contains templates, the it gets displayed. I'm not sure how it's worse from what we have now.

Do you know if Parsoid already has an API? Even if it does, I'm not sure how we would get the relevant portion of a huge description template (and there's more than one kind of template we would have to parse).

Patch-set: 1
Label: Verified=0