Update patch set 5
Patch Set 5:

Well going forward I believe we should match the desktop id's and classes. If these are inconsistent we should get them changed in desktop.

Ideally someone should be able to write a skin for desktop that works on mobile and existing skins like Vector should work to some capacity. This is what we should be aiming to get to.

I think we gain more by following patterns on desktop where acceptable.

Patch-set: 5
Label: Verified=0