Update patch set 1
Patch Set 1:

I can imagine this will be *very* useful. I don't think we've done it before, but we have quite a few hooks that we expose - this one probably being the most useful to other extensions, etc, so I think it would be good for us to start documenting these. Perhaps in a README.hooks or something like that?

Patch-set: 1