Update patch set 3
Patch Set 3:

The MinervaTemplate thing seems orthogonal to the rest of this commit, perhaps it would be better to split up this commit. Also, does it make sense to include Minerva* in MobileFrontend? My instinct is to put a desktop-specific skin in a completely separate extension expressly for that purpose. MobileFrontend is already quite large and complex, and I'm not sure we should be adding more to it we do not definitely need, especially something that is contextually different (a desktop skin).

Patch-set: 3