Update patch set 2
Patch Set 2: Code-Review-1

I dont think this is the best approach as what you're expecting to be set for $wgNearbyHost to be just the host portin of the URL, AND you're expecting the result of getApiUrl() to be the same for in either case, which is not necessarily true. If you have $wgNearbyHost set, you should probably just specify the entire URL for the API and use that instead of trying to munge two things togther that don't necessarily fit together.

Patch-set: 2
Label: Code-Review=-1