Update patch set 1
Patch Set 1:

you may want to rely on $wgCookiePrefix in the cookie name to make it easier to identify this cookie on different domains. it's a convention used in MW (though not necessarily all the time, particularly when you want a cookie available across different domains or subdomains, like stopMobileRedireect).

Patch-set: 1