Update patch set 7
Patch Set 7:

Desktop uses jquery suggestions plugin. We could use this but I've always felt something like this is overkill.

That said I've always assumed our search for a search term 'San' would be able to match 'Comic Sans' as well as 'San Francisco' but it seems in fact it doesn't so we could actually rewrite our highlighter to do a similar thing to jquery suggestions and make it highlight only terms at the start rather than terms anywhere - this simplifies the problem greatly..

Patch-set: 7