Do not display desktop view on mobile domain
I initially attempted to do this a less hacky way by attempting to more
accurately determine whether or not a user was on a mobile-specific
domain, but unfortunately it was an untennable solution since $wgServer
will always evaluate to the desktop domain. This solution should work,
although it will cause problems if there is a separate desktop and
mobile domain, X-Device gets set by proxy or webserver that handles desktop
domain, and the user attempts to access desktop domain. This seems like
an unlikely scenario (why would you ever set X-Device for a desktop-only
domain?). Also, I would've liked to have added tests for this, but
unfortunately there's no reasonable way to handle cookies in phpunit :(

Bug: 46473
Change-Id: I56f2c8a2626042bf7b9a1d94fbf14ec8e2f089c5