Update patch set 1
Patch Set 1:

From a user perspective this would be extremely useful. It gives a sense of the corresponding talk page and spikes curiosity...

How can be make this acceptable from a performance perspective?

Could we do something like PageImages but for sections and maybe save an entry on a save hook? It would also improve our MobileFormatter performance..

Patch-set: 1