Story 425: Nearby preview
Throwing massive curve ball.

Currently it is annoying on the nearby page to click an article link
then have to click back to get back to the article. This allows the
reader to preview all the articles from the comfort of the nearby
page and navigate away when ready.

The overlay that shows when clicking on an article shows the lead
section text and title with a link to the full article at the top
Any tables in the lead section are hidden

A user can right click/hold down the link to open it in a new
window if they don't like this behaviour

Move hacks into mobile.styles as these now apply to overlays in special
pages

Change-Id: Ifbebdb35cf5c6a758741a41822dd2361f4fa4ccb