Update patch set 1
Patch Set 1: Code-Review-1

This breaks nearby for me - that is, I no longer see it in my nav in alpha or beta. I tested in desktop chrome, and desktop chrome masquerading as an android 4.2 device.

When I check in console, it appears that my browser supports nearby:
> mw.mobileFrontend.supportsGeoLocation()
true

And it appears that the correct class is being added the html tag:
<html lang="en" dir="ltr" class="client-js supportsPositionFixed supports-geo-location transforms togglingEnabled">

Patch-set: 1
Reviewer: Awjrichards <91@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1