Update patch set 4
Patch Set 4: Verified+2 Code-Review+1

This is a nice approach! But I think this only partly solves the problem I think this story is attempting to solve - losing your place in the Nearby list when you back. If I continue to read the article in full, haven't changed my location, but go back to Nearby, a new request gets made to load the exact same content, and I'm taken back to the top of the content. I guess getting taken back to the top of the list is not a big deal, but having to fetch and load the same data twice (or more than that) seems silly to me.

It would be nice if we cached the Nearby data and only make new queries if the user has significantly changed location, but perhaps that's another story.

Patch-set: 4
Label: Verified=+2
Label: Code-Review=+1