Update patch set 4
Patch Set 4:

Maryana: Also the more I think about it - I actually think this is a good thing. If someone is interested in playing the wikipedia nearby photo hunt game they will click through. I think this would improve quality of photos that come through. We could always enable it on preview with a different funnel to run some tests on the quality of results..

Patch-set: 4