Update patch set 4
Patch Set 4:

Arthur: I actually think you're less likely to click back with this type of view. I imagine in the current setup people are clicking back when either
1) they discover they cannot upload a photo on the current page
2) they want to see what else is nearby.

I think the summary mode makes mode 1 and 2 easier.

That said it would be trivial to add a bit of caching as well on top of this and should be done as well (I think the preview is the ideal behaviour for this type of screen though!!).

Patch-set: 4