Update patch set 5
Patch Set 5: Code-Review-1

I have $wgResourceLoaderDebug = true;

And _everything_ works without this commit. I won't +1 a patchset that breaks Resource Loader, even if it breaks it only in debug mode.

We can add a photo upload button, but it might be tricky because of the not-up-to-date wgTitle.

Also, Maryana and I think that the title of the article doesn't belong to the Chrome, mainly because it can be really long (like 3 lines long).

Patch-set: 5
Label: Code-Review=-1