Update patch set 5
Patch Set 5:

Juliusz... I'm not expecting you to +1 that something that doesn't work for you but I'm trying to clarify if it works for you /without/ debug=true - if so it would be good to know what scripts are loading in what order to get a sense of the dependency tree that is being incorrectly built.

This works for me with debug=true and without so it's a little confusing.

Patch-set: 5