Update patch set 5
Patch Set 5:

"If someone is interested in playing the wikipedia nearby photo hunt game they will click through" – I'm not sure they will. It's not immediately clear that this is a preview, since it loads a pretty big chunk of the page and doesn't alter the UI in any significant way. It you're not looking carefully (which you probably wouldn't be if you were just casually hopping around all the nearby articles) it looks just like a regular Wikipedia article. Sure, if you're one of the limited number of users who knows the upload feature is there and will go hunting for it, you'd find it eventually, but that kills the joyous serendipity of landing on a page with no images and being asked to add one for the vastly larger number of logged in users who have never uploaded before.

If we want to build a preview feature, I think we should a) make something generic, so we can reuse it on other workflows like random article, and b) put a lot of care into crafting a solid UX/UI. Previews are tricky because they can easily slip into giving too much information or not enough. There's unresolved questions here, like how a preview could/would handle really short stubs, and what features/calls to action it should preserve or hide. Can you watchlist a preview? Can you edit/upload to one? Why (potentially) one and not the other? 

All of these things are still pretty unclear to me, and with a feature like this I think the devil is not in the technical implementation (which, as you've proven, can get hacked together fairly quickly/easily) but precisely in the UI details, which need more time to get hashed out.

Patch-set: 5
Label: Verified=0