Update patch set 3
Patch Set 3: Code-Review-2

(1 comment)

Code looks fine but 2 problemos... #1 is quite serious hence the -2 (this is the only reason for the -2)

1) Note that in alpha with hijacked links still working this change breaks the  CC-By-SA license agreement in that lazy loaded pages do not link to the history page. https://gerrit.wikimedia.org/r/#/c/59785/ disables link hijacking temporarily so I'd wait for that to be merged before merging this so we can obey the license.

Also I assume you have checked this with Vibha but with a watermark logo the footer looks extremely cluttered. Previously the watermark was in the top left to avoid this. Was this change intended? http://imgur.com/a9sNrQX

Patch-set: 3