Update patch set 5
Patch Set 5:

Since the nearby page fully relies on javascript, I've been thinking of the nearby page as a mini app inside the site.

Bear in mind when promoting the nearby feature we can always not promote the preview. I'd really encourage you to play with this some more with live data in the wild. Usually the previews are very small and only show 1 or 2 lines so are quite obviously not the full article (we don't show sub headings for sections or any tables). If you'd prefer I can stick this in alpha for the time being.

I'd encourage us to think about the user experience rather than the goal of creating more editors. I believe an experience without a preview even with some form of results caching is not going to be as smooth and fun to use (for instance we cannot cache page images so every time a user hits back there will be a delay as the images to the left of titles reload). I believe by making this feature a wonderful web app in itself we will do us much more long term benefit.

As I've also pointed out before there is no reason we cannot add the photo upload button to these previews in future.

Patch-set: 5