Update patch set 5
Patch Set 5:

Also you forget the add image to this page CTA only shows to logged in users. Non-logged in users (who can still use nearby) will not see it. This is the majority of our audience. (Thus why I say people that are aware of this feature will still click through)

Patch-set: 5