Update patch set 8
Patch Set 8:

I was hoping just to get the functionality in there and then get a designer to look at it and maybe turn it into a story for beta... after all if no one uses it it's probably not worth much more effort.

On that basis are you happy to +1/2?

Patch-set: 8