Update patch set 9
Patch Set 9:

please fix up yourself so I can merge tomorrow. With reference to ids - it's just a fake I'd. It's main purpose is uniqueness so it can be opened when you click on the heading. Unfortunately ids are not tied to number of h2s in a page. The nth section might have id n but it might also be n*2 - it all depends how many h3/4/5s came before it thus total sections does not work. Use Math.random or take the highest section id + number of new sections if you prefer the only real purpose of this is to look it up and it will only persist in the users current session but it must must must be unique. I wanted to keep this as simple as possible.

Patch-set: 9
Label: Verified=0