Update patch set 4
Patch Set 4:

All this does in dynamically load sub sections. Inspect the DOM for #content_1 and you will see it is empty but has a data attribute set. The purpose of this is to minimise initial payload and inject into the DOM when the sections are toggled.

Currently the link hijacking code is disabled in anticipation that that will refactored later along with the random code. If you prefer we could remove these all together and rewrite from scratch.

Patch-set: 4
Label: Verified=0