Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

It looks to me like setting $wgMFLeadPhotoUploadCssSelector = '*'; won't necessarily work as expected since its length would be non-0. Maybe either set this to '' or false to disable (using 'false' would be more consistent with how we do this elsewhere).

It would be nice to see a test for that scenario (when lead photo upload is disabled in config).

Also, see inline comment.

Patch-set: 1