Update patch set 1
Patch Set 1:

(1 comment)

"It looks to me like setting $wgMFLeadPhotoUploadCssSelector = '*'; won't necessarily work as expected since its length would be non-0. Maybe either set this to '' or false to disable (using 'false' would be more consistent with how we do this elsewhere).
It would be nice to see a test for that scenario (when lead photo upload is disabled in config).

I don't really understand what you mean here 'wgMFLeadPhotoUploadCssSelector' needs to be a string and the * example was only a suggestion and we are unlikely to ever use in production.

$container.find( mw.config.get( 'wgMFLeadPhotoUploadCssSelector' ) ).length === 0;

needsPhoto because of that above line will only return true if no elements match the css selector.

There is a slight edge case where if a page has no elements  - not even a heading (very very unlikely) a page will need a photo.

This seems like premature optimisation as we are currently not planning on disabling anywhere.

I'm not sure how an additional test will help here...

Patch-set: 1