Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

@jdlrobson in reply to your reply to my comment:
to the first point, you're right; in haste i misread the javascript. but given what you've said:
"'wgMFLeadPhotoUploadCssSelector' needs to be a string and the * example was only a suggestion and we are unlikely to ever use in production." then I think the comment in MobileFrontend.php for this config var should be made clearer to reflect this reality. Would an empty string be acceptable? If so, I would prefer that be the suggested use case since it feels more intuitive/that's more common in MW. If not, then it's fine I guess but regardless I think the documentation should be made more clear.

In regards to the testing, I didn't see a test scenario for when wgMFLeadPhotoUploadCssSelector is set to something that will 'turn off lead photo uploads altogether', though it seems to me if we support that behavior, it's something that should be tested as well.

Patch-set: 1
Label: Code-Review=-1