Update patch set 1
Patch Set 1:

Just to clarify -- apparently the links about license at the bottom of the page are rendered by the mobile skin. When in Zero, if the link leads to outside of zero, we need to show the warning about potential charges, or, in the ideal case, make so that all the legal pages reside on the same site as zero page.  Your suggestions are welcome.

Patch-set: 1
Reviewer: Yurik <571@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0