Update patch set 2
Patch Set 2: Code-Review-1

what is the advantage to this approach? i still dont totally understand why, but using the ResourceLoaderRegisterModules is discouraged: https://www.mediawiki.org/wiki/Manual:Hooks/ResourceLoaderRegisterModules

If the goal of this is to cleanup MobileFrontend.php, why not just stick the $wgResourceModules into a separate file and include() it in MobileFrontend.php? That would also be simpler/lighter weight.

Patch-set: 2
Label: Code-Review=-1