Update patch set 6
Patch Set 6: Verified+2 Code-Review+1

Clearly there 999 ways to skin this cat. I don't see the advantage to keeping the definitions of $wgResourceModules out of the global scope, nor do I really see the advantage to having a separate file return an array. My hunch is things will perform just a little bit faster if we keep it all in the global scope and include a separate file (although Krinkle's point about the overhead of require_once() vs using a setup hook is well taken). I'd be curious to performance test the different approaches and pick the fastest option. Unfortunately, I need to go so I can't do this right now, and I wont be back to work til next Monday.

So: I've amended the latest patchset to remove an unneeded line. I think this patchset is fine to merge as it currently is, but if someone wants to take some time and perf test the different possible approaches, that would be really cool and would give us a data-driven decision as opposed to a 'I just like this way better than the other'

Patch-set: 6
Reviewer: Awjrichards <91@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=+2
Label: Code-Review=+1